feature: add truncation support (#7)
* Document title in window title and other fixes

This adds the document title to the window title. It was thought that perhaps this would be viable in iTerm in macOS with the touch-bar thing, as it displays the window title there, but it doesn't update the title there when it changes.

This also fixes an issue that viewing the outline could cause the document to show the wrong number of words.

* Update version.

* This adds truncation support.

Truncation is enabled via the configuration file, and deletes
whole words with a single press of the backspace key.